await/async on intializing projects from database
