+++ backend/index.js
@@ -2,7 +2,7 @@ const express = require('express')
 const app = express()
 const port = 3001
 
-const merchant_model = require('./merchantModel')
+const project_model = require('./projectModel')
 
 app.use(express.json())
 app.use(function (req, res, next) {
@@ -13,7 +13,7 @@ app.use(function (req, res, next) {
 });
 
 app.get('/', (req, res) => {
-  merchant_model.getProjects()
+  project_model.getProjects()
   .then(response => {
     res.status(200).send(response);
   })
@@ -21,38 +21,6 @@ app.get('/', (req, res) => {
     res.status(500).send(error);
   })
 })
-
-app.post('/merchants', (req, res) => {
-  merchant_model.createMerchant(req.body)
-  .then(response => {
-    res.status(200).send(response);
-  })
-  .catch(error => {
-    res.status(500).send(error);
-  })
-})
-
-app.delete('/merchants/:id', (req, res) => {
-  merchant_model.deleteMerchant(req.params.id)
-  .then(response => {
-    res.status(200).send(response);
-  })
-  .catch(error => {
-    res.status(500).send(error);
-  })
-})
-app.put("/merchants/:id", (req, res) => {
-  const id = req.params.id;
-  const body = req.body;
-  merchant_model
-    .updateMerchant(id, body)
-    .then((response) => {
-      res.status(200).send(response);
-    })
-    .catch((error) => {
-      res.status(500).send(error);
-    });
-});
 
 app.listen(port, () => {
   console.log(`App running on port ${port}.`)
