+++ frontend/src/pages/ProjectIndex.tsx
@@ -3,7 +3,6 @@ import React, { useState, useEffect } from 'react'
 import { createStyles} from '@mantine/core'
 import { PreviewCarousel } from '../components/index';
 import { Outlet} from 'react-router-dom';
-import type { Projects } from '../projects';
 
 export interface merchants {
   [key: string]: { name: string, email: string };
