+++ frontend/src/projects.ts
@@ -6,36 +6,36 @@ export function saveProjects(projects: Projects): void {
     return localStorage.setItem(PROJECTS_KEY, JSON.stringify(projects));
 }
 
-function initializeProjects(): Projects {
-    let projectJSON: string[] = [];
-    let collection: Projects | null = null;
-    
-    fetch('http://localhost:3001')
-    .then(response => {
-      return response.text();
-    })
-    .then(data => {
-      projectJSON = JSON.parse(data)
-      for (var project of projectJSON)
-      {
-
-        collection[project.id] = {previewImage: project.previewimage, title: project.title, images: project.images.split(','), data: project.data, dataInfo: project.datainfo}
-      }
-    });
-    
-    collection = collection ? collection : backup
+async function initializeProjects(): Promise<Projects> {
+    let collection: Projects = {};
+    let response: Response;
+    console.log("fetch")
+    try {
+      response = await fetch("http://localhost:3001")
+    } catch (e) {
+      console.log('Unable to fetch from DB, using backup')
+      return backup
+    }
+    const projects = await response.json()
+    for (var project of projects)
+    {
+      collection[project.id] = {previewImage: project.previewimage, title: project.title, images: project.images.split(','), data: project.data, dataInfo: project.datainfo}
+    }
     saveProjects(collection)
     return collection
 }
 
-export function getProjects(): Projects {
+export async function getProjects(): Promise<Projects> {
     let projects: Projects | null = null;
+    
     try {
         // @ts-expect-error OK to throw here since we're catching
         projects = JSON.parse(localStorage.getItem(TODOS_KEY));
-      } catch (e) {}
-      if (!projects) {
-        projects = initializeProjects();
+      } catch (e) {
+      }
+      
+      if (!projects) {  
+        projects = await initializeProjects();
       }
     return projects;
 }
